task .102

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@ int main(void)
 {
 	int i = 3, j = 4;
 
-	unsigned int sum = 0, pp = 2, p = 3;
+	unsigned long sum = 0, pp = 2, p = 3;
 
 	printf("1, 2, ");
